pyproject.toml
[project]
name = "silkroad"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "hades", email = "nischalbhattaraipi@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "alpaca-py>=0.43.2",
    "arcticdb>=6.3.2",
    "backtrader>=1.9.78.123",
    "cvxpy>=1.7.3",
    "fastapi[standard]>=0.123.0",
    "flax>=0.12.0",
    "fredapi>=0.5.2",
    "jax[cuda]>=0.8.0",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "pandas-market-calendars>=5.1.3",
    "plotly>=6.5.0",
    "pydantic>=2.12.4",
    "rich>=14.2.0",
    "riskfolio-lib>=7.0.1",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.5",
    "sympy>=1.14.0",
    "ta-lib>=0.6.8",
    "tqdm>=4.67.1",
    "xgboost>=3.1.1",
    "yfinance>=0.2.66",
]

[project.scripts]
silkroad = "silkroad.app.main:main"

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "nbformat>=5.10.4",
    "pytest>=9.0.1",
]

[tool.ty.rules]
index-out-of-bounds = "ignore"
